# [RTG][Elaboration] Move ConstantLike check after TypeSwitch for better performance