--- a/lib/Dialect/RTG/Transforms/ElaborationPass.cpp
+++ b/lib/Dialect/RTG/Transforms/ElaborationPass.cpp
@@ -696,6 +696,28 @@ public:
   }
 
   FailureOr<DeletionKind> visitExternalOp(Operation *op) {
+    if (op->hasTrait<OpTrait::ConstantLike>()) {
+      SmallVector<OpFoldResult, 1> result;
+      auto foldResult = op->fold(result);
+      (void)foldResult; // Make sure there is a user when assertions are off.
+      assert(succeeded(foldResult) &&
+             "constant folder of a constant-like must always succeed");
+      auto attr = dyn_cast<TypedAttr>(result[0].dyn_cast<Attribute>());
+      if (!attr)
+        return op->emitError(
+            "only typed attributes supported for constant-like operations");
+
+      auto intAttr = dyn_cast<IntegerAttr>(attr);
+      if (intAttr && isa<IndexType>(attr.getType()))
+        internalizeResult<IndexValue>(op->getResult(0), intAttr.getInt());
+      else if (intAttr && intAttr.getType().isSignlessInteger(1))
+        internalizeResult<BoolValue>(op->getResult(0), intAttr.getInt());
+      else
+        internalizeResult<AttributeValue>(op->getResult(0), attr);
+
+      return DeletionKind::Delete;
+    }
+
     // TODO: we only have this to be able to write tests for this pass without
     // having to add support for more operations for now, so it should be
     // removed once it is not necessary anymore for writing tests
@@ -970,28 +992,6 @@ public:
   }
 
   FailureOr<DeletionKind> dispatchOpVisitor(Operation *op) {
-    if (op->hasTrait<OpTrait::ConstantLike>()) {
-      SmallVector<OpFoldResult, 1> result;
-      auto foldResult = op->fold(result);
-      (void)foldResult; // Make sure there is a user when assertions are off.
-      assert(succeeded(foldResult) &&
-             "constant folder of a constant-like must always succeed");
-      auto attr = dyn_cast<TypedAttr>(result[0].dyn_cast<Attribute>());
-      if (!attr)
-        return op->emitError(
-            "only typed attributes supported for constant-like operations");
-
-      auto intAttr = dyn_cast<IntegerAttr>(attr);
-      if (intAttr && isa<IndexType>(attr.getType()))
-        internalizeResult<IndexValue>(op->getResult(0), intAttr.getInt());
-      else if (intAttr && intAttr.getType().isSignlessInteger(1))
-        internalizeResult<BoolValue>(op->getResult(0), intAttr.getInt());
-      else
-        internalizeResult<AttributeValue>(op->getResult(0), attr);
-
-      return DeletionKind::Delete;
-    }
-
     return TypeSwitch<Operation *, FailureOr<DeletionKind>>(op)
         .Case<
             // Index ops